+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/generation/ParadoxLocalisationGenerationChooser.kt
@@ -43,8 +43,8 @@ class ParadoxLocalisationGenerationChooser(
         myOptionControls = emptyArray()
 
         return panel {
-            row {
-                panel { configureSouthOptionsPanel() }
+            panel {
+                configureOptionsPanel()
             }
             separator()
             row {
@@ -53,19 +53,13 @@ class ParadoxLocalisationGenerationChooser(
         }
     }
 
-    private fun Panel.configureSouthOptionsPanel() {
+    private fun Panel.configureOptionsPanel() {
+        collapsibleGroup(PlsBundle.message("settings.generation")) { configureOptionsGroup() }
+    }
+
+    private fun Panel.configureOptionsGroup() {
         val settings = PlsSettings.getInstance().state.generation
 
-        // moveIntoLocalisationGroups
-        row {
-            checkBox(PlsBundle.message("settings.generation.moveIntoLocalisationGroups"))
-                .bindSelected(settings::moveInfoLocalisationGroups.toAtomicProperty())
-        }
-        // newLineBetweenLocalisationGroups
-        row {
-            checkBox(PlsBundle.message("settings.generation.newLineBetweenLocalisationGroups"))
-                .bindSelected(settings::newLineBetweenLocalisationGroups.toAtomicProperty())
-        }
         // localisationStrategy
         row {
             val property = AtomicProperty(settings.localisationStrategy)
@@ -77,6 +71,16 @@ class ParadoxLocalisationGenerationChooser(
                 .visibleIf(property.transform { it == LocalisationGeneration.SpecificText })
             localeComboBox(withAuto = true).bindItem(settings::localisationStrategyLocale.toAtomicProperty(ParadoxLocaleManager.ID_AUTO))
                 .visibleIf(property.transform { it == LocalisationGeneration.FromLocale })
+        }
+        // newLineBetweenLocalisationGroups
+        row {
+            checkBox(PlsBundle.message("settings.generation.newLineBetweenLocalisationGroups"))
+                .bindSelected(settings::newLineBetweenLocalisationGroups.toAtomicProperty())
+        }
+        // moveIntoLocalisationGroups
+        row {
+            checkBox(PlsBundle.message("settings.generation.moveIntoLocalisationGroups"))
+                .bindSelected(settings::moveIntoLocalisationGroups.toAtomicProperty())
         }
     }
 
@@ -94,18 +98,21 @@ class ParadoxLocalisationGenerationChooser(
 
     private fun createSelectAllAction(): SelectAction {
         return SelectAction(PlsBundle.message("generation.localisation.select.all")) {
+            myTree.clearSelection() // clear first to keep in correct order
             selectElements(myElements)
         }
     }
 
     private fun createSelectMissingAction(): SelectAction {
         return SelectAction(PlsBundle.message("generation.localisation.select.missing")) {
+            myTree.clearSelection() // clear first to keep in correct order
             selectElements(myElements.filter { it.info.missing }.toTypedArray())
         }
     }
 
     private fun createSelectMissingAndCheckedAction(): SelectAction {
         return SelectAction(PlsBundle.message("generation.localisation.select.missingAndChecked")) {
+            myTree.clearSelection() // clear first to keep in correct order
             selectElements(myElements.filter { it.info.missing && it.info.check }.toTypedArray())
         }
     }
@@ -113,7 +120,7 @@ class ParadoxLocalisationGenerationChooser(
     private fun createSelectNoneAction(): SelectAction {
         return SelectAction(PlsBundle.message("generation.localisation.select.none")) {
             myTree.clearSelection()
-            doOKAction()
+            // doOKAction() // do not quit dialog
         }
     }